updated for auth0
